requirements.txt
discord-py
python-dotenv
openai
python-dotenv
pydantic<2.0
pymongo
tiktoken
requests
text-generation
tokenizers
transformers
pillow
webuiapi
stability-sdk